physai: Add physai train subcommand
- pipeline.py: _generate_train_sbatch, run_train (syncs config, model config,
  verifies dataset, submits sbatch with container protocol)
- cli.py: train subparser with --config, --dataset, --model-config-root, --max-steps
- tests: test_generate_train_sbatch (35 total)
